// types/writingData.ts
export type TWritingData = {
  writingData: {
    fiction: {
      title: string;
      text: string;
    };
    scriptwriting: {
      title: string;
      text: string;
    };
    stage: {
      title: string;
      text: string;
    };
    nonfiction: {
      title: string;
      text: string;
    };
    journalism: {
      title: string;
      text: string;
    };
  };
};